pyproject.toml
[project]
name = "eurogames"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastcore>=1.8.1",
    "fastlite>=0.1.3",
    "flask>=3.1.0",
    "jinja2>=3.1.6",
    "python-fasthtml>=0.12.12",
    "requests>=2.31.0",
]
